backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
drf-yasg==1.21.11
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
pillow==12.0.0
pytest
pytest-django
pytest-cov
itsdangerous>=2.1.2
celery==5.4.0
redis==5.2.1
django-celery-beat==2.8.0
django-redis==5.4.0
flower==2.0.1
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0

# CV Analysis with LLaMA 2 (CPU-only, no GPU required)
numpy<2
pinecone-client==5.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-pinecone==0.2.0
pypdf==5.2.0
replicate==0.34.1
sentence-transformers==2.7.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.2.0+cpu
torchaudio==2.2.0+cpu

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0
cloudscraper==1.2.71
requests==2.32.3